small steps
the update to tailwind config is a possible fix for the tarot cards not scaling.  For some reason on build it wasn't seeing h-80, and therefor not including it when building.

Everything else this update was a bit of qol, a bit more progress on filling in of P5R, and some more infrastructure stuff